// src/interfaces/entries.ts
// Generated by https://quicktype.io

import { EntryStatus } from "./reducerActions";

export interface AllEntriesResponse {
    data: ApiEntry[];
}

export interface ApiEntry {
    _id:         string;
    description: string;
    createdAt:   number;
    status:      EntryStatus;
    __v:         number;
}